+++ web/src/components/molecules/Accessibility/types.ts
@@ -1,0 +1,8 @@
+export type PublicScope = "private" | "public"; // Add "limited" when functionality becomes available
+
+export type Model = {
+  id: string;
+  name?: string;
+  public: boolean;
+  key?: string;
+};
