+++ web/src/components/organisms/CMSWrapper/hooks.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 import { useParams, useLocation, useNavigate } from "react-router-dom";
 
 import Notification from "@reearth-cms/components/atoms/Notification";
-import { PublicScope } from "@reearth-cms/components/molecules/Accessibility";
+import type { PublicScope } from "@reearth-cms/components/molecules/Accessibility/types";
 import { Role } from "@reearth-cms/components/molecules/Workspace/types";
 import {
   useCreateWorkspaceMutation,
